+++ app/assets/javascripts/govuk_publishing_components/analytics-ga4/ga4-event-tracker.js
@@ -84,7 +84,7 @@ window.GOVUK.Modules = window.GOVUK.Modules || {};
         }
       }
 
-      window.dataLayer.push(schema)
+      window.GOVUK.analyticsGA4.core.sendData(schema)
     }
   }